# lib/api-zod/src/generated/types/paymentStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentStatusStatus } from "./paymentStatusStatus";

export interface PaymentStatus {
  checkoutRequestId: string;
  status: PaymentStatusStatus;
  resultCode?: string | null;
  resultDesc?: string | null;
  amount?: number | null;
  phone?: string | null;
  offerName?: string | null;
  recipientPhone?: string | null;
  createdAt?: string | null;
}