# lib/api-zod/src/generated/types/stkPushRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StkPushRequest {
  /** Payer phone number — receives STK push (254XXXXXXXXX) */
  phone: string;
  /** Recipient phone number — who receives the bundle (if different from payer) */
  recipientPhone?: string | null;
  /** The offer ID to purchase */
  offerId: string;
  /** Amount to pay */
  amount: number;
}